Move shared files
Node.js is very picky about files not being outside the root of the module but vite doesn't care so I moved the shared files into the server folder.
