+++ client/src/components/Job.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { inject, computed } from 'vue';
 import type { IX1Client } from '@/plugins/IX1Client';
-import { Stage } from "@/../../shared/src/X1Messages";
+import { Stage } from "../../../server/src/shared/X1Messages";
 
 const x1Client = inject<IX1Client>("x1Client");
 if (x1Client === undefined)
