+++ server/src/shared/AmsTypes.ts
