+++ server/src/shared/LogLevel.ts
@@ -1,0 +1,8 @@
+export enum LogLevel
+{
+  Error       = 0,
+  Warning     = 1,
+  Information = 2,
+  Debug       = 3,
+  Trace       = 4
+}
